+++ frykit/shp/mask.py
@@ -8,7 +8,7 @@ from numpy.typing import ArrayLike, NDArray
 
 from frykit.calc import asarrays, is_monotonic_decreasing, is_monotonic_increasing
 from frykit.shp.typing import PolygonType
-from frykit.utils import format_type_error
+from frykit.utils import deprecator, format_type_error
 
 
 def polygon_mask(
@@ -217,3 +217,8 @@ def polygon_mask2(
         mask = mask[::-1, :]
 
     return mask
+
+
+@deprecator(alternative=polygon_mask)
+def polygon_to_mask(polygon: PolygonType, x: ArrayLike, y: ArrayLike) -> NDArray:
+    return polygon_mask(polygon, x, y, include_boundary=False)